requirements.txt
pygame>=2.5.0
numpy>=1.24.0
cupy-cuda12x>=12.0.0
pytest>=7.4.0
pyyaml>=6.0.1  # for config/display.yml